Fix Safari iframes redirect
Recently changes in Safari broke the existing iframes fix found in
CoalescingPanda::ControllerHelpers#session_check. Fixed the logic and
moved it to the authorization method to ensure it is only performed
after authorization is performed. Added a warning log message to the
#session_check method to notify users it has been deprecated.

Bumped version to 4.6.0.rc1 for testing before releasing final version.

refs: PFS-15006
Change-Id: I6b732e75b8eed9fff8bbe10bf887ffbeceac8b67